Split signals diagnostic dump into its own module
The diagnostic-corpus json dump is a cluster of free functions unrelated
to the extraction service's state. Move it to a sibling signals/dump.rs so
service.rs is only the extraction service, and keep the dump helpers and
their tests private to the new module.
