+++ bae-core/src/signals/mod.rs
@@ -31,6 +31,8 @@ pub mod barcode;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub mod disc_id;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod dump;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
 mod fast_pass;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub mod origin;
