+++ bae-core/src/signals/service/tests.rs
@@ -617,42 +617,3 @@ async fn three_starts_cancel_each_predecessor() {
         "expected a Settled snapshot from the completed run",
     );
 }
-
-#[test]
-fn sanitize_short_key_passes_through() {
-    assert_eq!(
-        sanitize_key_for_filename("abc_def-123"),
-        "abc_def-123".to_string(),
-    );
-}
-
-#[test]
-fn sanitize_replaces_unsafe_characters() {
-    let out = sanitize_key_for_filename("/a/b.c d#e");
-    assert!(out
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || c == '-' || c == '_'));
-}
-
-#[test]
-fn sanitize_long_key_truncated_with_hash_suffix() {
-    // Pathological: 2000-char input. Result must fit the cap and two
-    // different long inputs must produce distinct filenames.
-    let long_a: String = "a".repeat(2000);
-    let mut long_b = String::from("b");
-    long_b.push_str(&"a".repeat(1999));
-
-    let out_a = sanitize_key_for_filename(&long_a);
-    let out_b = sanitize_key_for_filename(&long_b);
-
-    assert!(
-        out_a.len() <= SANITIZED_FILENAME_CAP,
-        "sanitized length {} exceeds cap {SANITIZED_FILENAME_CAP}",
-        out_a.len(),
-    );
-    assert_ne!(
-        out_a, out_b,
-        "distinct long keys must map to distinct filenames"
-    );
-    assert!(out_a.ends_with(|c: char| c.is_ascii_hexdigit()));
-}
